Allow different sort field/order with shortcode
Added the ability to specify a sort field and order when using the shortcode, e.g.

[wp-publication-archive categories="newsletters" sort="title" order="asc" /]

This would return all publications within the newsletters category sorted ascending by post title, i.e. alphabetically.

I hard coded the use of 'date' and 'title' to be converted to 'post_date' and 'post_title' as I wasn't sure of the security implications of allowing users to pass the database field name directly.